E commerce/requirement.txt
flask
mysql-connector-python
flask-bcrypt